collect all login functions inside one controller
