+++ server/src/controller/auth.js
@@ -4,34 +4,33 @@ const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 
 const { checkEmail } = require('../queries');
+const loginSchema = require('./validation/loginSchema');
 
-// eslint-disable-next-line no-unused-vars
-const checkUserEmail = async (req, res, next) => {
-  const { email } = req.body;
-  const [users] = await checkEmail(email);
-  if (users) {
-    const { dataValues } = users;
-    req.userData = dataValues;
-    next();
-  } else {
-    throw Boom.unauthorized('Email does not exist, signup first');
-  }
-};
-
-const signIn = async (req, res, next) => {
-  const hashedPassword = req.userData.password;
-  const { password } = req.body;
+const login = async (req, res, next) => {
+  const { email, password } = req.body;
   try {
-    const result = await bcrypt.compare(password, hashedPassword);
-    if (result === false) {
-      throw Boom.unauthorized('invalid password');
+    const { error } = await loginSchema.validate(req.body, {
+      abortEarly: false,
+    });
+    if (error) throw Boom.badRequest(error.details.map((e) => e.message));
+    const [users] = await checkEmail(email);
+    if (users) {
+      const { dataValues } = users;
+      const hashedPassword = dataValues.password;
+      const result = await bcrypt.compare(password, hashedPassword);
+      if (result === false) {
+        throw Boom.unauthorized('invalid password');
+      } else {
+        const token = jwt.sign({ id: dataValues.id }, process.env.SECRET_KEY);
+        res.cookie('token', token).status(200);
+        res.json({ msg: 'logged in successfully' });
+      }
     } else {
-      const token = jwt.sign({ id: req.userData.id }, process.env.SECRET_KEY);
-      res.cookie('token', token).status(200);
-      res.json({ msg: 'logged in successfully' });
+      throw Boom.unauthorized('Email does not exist, signup first');
     }
   } catch (error) {
     next(error);
   }
 };
-module.exports = { checkUserEmail, signIn };
+
+module.exports = login;
